refactor: remove `utils.ts`
- move string functions to `helpers/string`
- replaced `deprecated` util with `warnOnce` from `helpers/deprecate`
