+++ packages/core/src/features/st-var.ts
@@ -8,14 +8,13 @@ import type { EvalValueData, EvalValueResult } from '../functions';
 import { isChildOfAtRule } from '../helpers/rule';
 import { walkSelector } from '../helpers/selector';
 import { stringifyFunction, getStringValue, strategies } from '../helpers/value';
+import { stripQuotation } from '../helpers/string';
 import { ignoreDeprecationWarn } from '../helpers/deprecation';
 import type { ImmutablePseudoClass, PseudoClass } from '@tokey/css-selector-parser';
 import type * as postcss from 'postcss';
 import { processDeclarationFunctions } from '../process-declaration-functions';
 import { Diagnostics } from '../diagnostics';
 import { unbox } from '../custom-values';
-// ToDo: move
-import { stripQuotation } from '../utils';
 import type { ParsedValue } from '../types';
 
 export interface VarSymbol {
